+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - GLS-173 - Add model + endpoint for handling UK total trade data by country
 - GLS-186 - Dataservices addition for market trends
 - GLS-186 - Dataservices addition for trade highlights
-- GLS-208 - Ingest market guides data from Data Workspace
+- GLS-232 - Ingest ONS data for UK total trade from Data Workspace
+- GLS-234 - Ingest ONS data for UK trade in services from Data Workspace
 
 ### Bugs fixed